// src/app/types/raza.types.ts
/**
 * @fileoverview Definición de tipos para la entidad Raza
 * @module types/raza.types
 */

/**
 * Interfaz que representa una Raza en el sistema
 * @interface Raza
 */
export interface Raza {
  /**
   * Identificador único de la raza
   */
  Valor: number;

  /**
   * Descripción de la raza
   */
  Descripcion: string;
}